Add rates script for utility dataclass and functions
