+++ pyproject.toml
@@ -121,6 +121,8 @@ ignore = [
     "E501",
     # DoNotAssignLambda
     "E731",
+    # Redefinition of unused import
+    "F811",
 ]
 
 [tool.ruff.lint.per-file-ignores]
