+++ rate_design_platform/second_pass.py
@@ -13,54 +13,15 @@ import pandas as pd
 from ochre import Dwelling  # type: ignore[import-untyped]
 from ochre.utils import default_input_path  # type: ignore[import-untyped]
 
-from rate_design_platform.Analysis import (  # type: ignore[import-unresolved]
-    MonthlyResults,
-    SimulationResults,
+from rate_design_platform.Analysis import MonthlyResults, SimulationResults
+from rate_design_platform.utils.rates import (
     TOUParameters,
+    calculate_monthly_intervals,
 )
 
 # Define constants
 seconds_per_hour = 3600
 hours_per_day = 24
-
-
-def calculate_monthly_intervals(start_time: datetime, end_time: datetime, time_step: timedelta) -> list[int]:
-    """
-    Calculate number of time_step-long intervals in a given month
-
-    Args:
-        start_time: Start time of the simulation
-        end_time: End time of the simulation
-        time_step: Time step of the simulation
-
-    Returns:
-        Number of time_step-long intervals intervals for each month from start_time to end_time
-    """
-    intervals = []
-    current_time = start_time
-    while current_time < end_time:
-        # Calculate start and end of the current month within our time range
-        month_start = current_time
-        # Find the start of the next month
-        if current_time.month == 12:
-            next_month_start = current_time.replace(
-                year=current_time.year + 1, month=1, day=1, hour=0, minute=0, second=0, microsecond=0
-            )
-        else:
-            next_month_start = current_time.replace(
-                month=current_time.month + 1, day=1, hour=0, minute=0, second=0, microsecond=0
-            )
-        month_end = min(next_month_start, end_time)
-        # Calculate the duration of this month (within our time range)
-        month_duration = month_end - month_start
-        # Calculate number of intervals in this month
-        num_intervals = int(month_duration.total_seconds() / time_step.total_seconds())
-        intervals.append(num_intervals)
-
-        # Move to next month
-        current_time = next_month_start
-
-    return intervals
 
 
 def calculate_monthly_bill(simulation_results: SimulationResults, rates: list[np.ndarray]) -> list[float]:
